use generate-id(.) for bnode-id
xdmp:random will not work as the bnode-id needs to be taken in
different places (subject, object etc.) and calling xdmp:random will
generate two different values.

The rdfa.xqy solution used counting of preceding nodes which is not the
best way to do this, generate-id(.) should work just as fine and it is
less memory hungry.